system/requirements.txt
# Usage: pip install -r requirements.txt

py-cpuinfo
psutil
Flask